Separate class into two files
